Fixed 386 build (#2591)
* Fixed 386 build again

+++ util/types/time.go
@@ -1767,14 +1767,14 @@ func (t Time) convertDateFormat(b rune, buf *bytes.Buffer) error {
 	case 'X':
 		year, _ := t.Time.YearWeek(2)
 		if year < 0 {
-			fmt.Fprintf(buf, "%v", math.MaxUint32)
+			fmt.Fprintf(buf, "%v", uint64(math.MaxUint32))
 		} else {
 			fmt.Fprintf(buf, "%04d", year)
 		}
 	case 'x':
 		year, _ := t.Time.YearWeek(3)
 		if year < 0 {
-			fmt.Fprintf(buf, "%v", math.MaxUint32)
+			fmt.Fprintf(buf, "%v", uint64(math.MaxUint32))
 		} else {
 			fmt.Fprintf(buf, "%04d", year)
 		}
